Move ProjectDescription component to be under DiscoverySidebar
