+++ app/assets/src/components/views/discovery/DiscoverySidebar/ProjectDescription.jsx
