+++ app/assets/src/components/views/discovery/DiscoverySidebar/discovery_sidebar.scss
