+++ app/assets/src/components/views/discovery/DiscoverySidebar/index.js
@@ -1,0 +1,3 @@
+import DiscoverySidebar from "./DiscoverySidebar";
+
+export default DiscoverySidebar;
