+++ app/assets/src/components/views/discovery/DiscoverySidebar/project_description.scss
